plic: clear irq pending when init plic
there maybe some irq pending triggered by uboot driver,
the pending will not be cleared even when irq source is
cleared. when the irq source driver devm_request_irq, the
isr will be triggered, but the devm_request_irq is not finished,
and the driver is not ready, there may be some problems.
so, try to clear all irq pending when init plic

Change-Id: I88db4db0af7f7e8a9893b0ffb9db07b7d0203c37